requirements.txt
Django==3.2.6
django-cors-headers==3.7.0
mysql-client==0.0.1
PyMySQL==1.0.2
bcrypt==3.2.0
PyJWT==2.1.0
requests==2.26.0
